backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# AWS & Storage
boto3==1.34.0
botocore==1.34.0

# Payment
requests==2.31.0

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0
